Added modes and counters to settings panel
Settings panel :: Temporary implementation of the settings panel. Eventually it will be
made using tkinter library for better looking panel.

Modes :: Different modes where added to the visualizer, such as map
creator, running, done and no solution modes. The will show in the
settings panel.

Counters :: Simple counter renderer for the A* pathfinding lengths of
the open, closed and path sets used in the algorithm. They will show in
the settings panel.